ci: run benchmarks on demand (#26)
The bench workflow[1] is both very slow (~4min per
run) and flaky[2].

Here we attempt to transform it into an on-demand
workflow, made easier by an automated comment with
instructions for kicking it off. The initial
transformation was made primarily by copy/pasting
the old workflow into https://claude.ai and
iterating.

[1]: https://github.com/mccutchen/websocket/actions/workflows/bench.yaml
[2]: https://github.com/mccutchen/websocket/actions/runs/12901137705/job/35972729251