refactor: complete rewrite of the frame type (#39)
With this experimental refactor, a decoded websocket frame is
represented in memory as a single header byte and a payload byte slice.

The single header byte gives us all the info we need at the application
level (opcode, fin bit, RSV bits).

Because we unmask the payload at read time and already require callers
to specify a mask when writing a frame, we do not need to store the
masked bit or the mask key in the frame itself. (This does, however,
require passing a mode into ReadFrame so that we can still reject
unmasked client frames.)

And because correctly constructing the header byte for a frame is tricky, a
new `NewFrame` constructor is now the only way to create a frame.