fix: implement full closing handshake (#63)
This somewhat messy change adds support for properly closing websocket
connections via the [closing handshake][1], where the peer that initiates a
close waits for a reply before actually closing the TCP connection.

This fixes https://github.com/mccutchen/websocket/issues/47, which notes
a mystery:

> It seems like the autobahn test suite should probably cover this, but
we're passing those tests with our incomplete implementation.

I now think that's because most/all of the closing scenarios tested by
autobahn are for edge cases where it's appropriate for the server to
immediately close the connection without waiting for a reply. As noted
in a comment below, see [section 7.1.7 of the RFC][2] and search for
"Fail the" in the RFC to see all the scenarios where immediately closing
the connection is the correct/recommended/required behavior.

AFAICT the "full" closing handshake is mostly going to be useful if/when
application code wants to signal errors.

[1]: https://datatracker.ietf.org/doc/html/rfc6455#section-1.4
[2]: https://datatracker.ietf.org/doc/html/rfc6455#section-7.1.7